+++ CheMPS2/include/chemps2/Problem.h
@@ -126,7 +126,11 @@ namespace CheMPS2{
          /** \param docc Array which contains for each irrep the number of doubly occupied orbitals
              \param sp_threshold Threshold to detect Delta_g and Delta_u partners based on single-particle energies */
          void setup_reorder_dinfh(int * docc, const double sp_threshold=1e-5);
-         
+
+         //! Check that ROHF-style occupancies are compatible with the currently targeted symmetry sector.
+         /** \param occupancies Array which contains per DMRG orbital the ROHF-style occupancy (0, 1 or 2) */
+         bool check_rohf_occ( int * occupancies );
+
       private:
       
          //Pointer to the Hamiltonian --> constructed and destructed outside of this class